tests and documentation
